+++ source/main/providers/dictionary-provider.js
@@ -15,8 +15,10 @@
 
 const EventEmitter = require('events')
 const NSpell = require('nspell')
+const path = require('path')
 const fs = require('fs')
-const ipc = require('electron').ipcMain
+
+const { ipcMain, app } = require('electron')
 const { getDictionaryFile } = require('../../common/lang/i18n.js')
 const { promisify } = require('util')
 const readFile = promisify(fs.readFile)
@@ -29,8 +31,12 @@ const readFile = promisify(fs.readFile)
 class DictionaryProvider extends EventEmitter {
   constructor () {
     super()
+    // Array containing all loaded NSpell dictionaries
     this._typos = []
-    this._loadedDicts = [] // Array containing the language codes for which checking currently works
+    // Array containing the language codes for which checking currently works
+    this._loadedDicts = []
+    // Path to the user dictionary
+    this._userDictionaryPath = path.join(app.getPath('userData'), 'user.dic')
 
     // Inject global methods
     global.dict = {
@@ -39,11 +45,13 @@ class DictionaryProvider extends EventEmitter {
     }
 
     // Listen for synchronous messages from the renderer process for typos.
-    ipc.on('typo', (event, message) => {
+    ipcMain.on('typo', (event, message) => {
       if (message.type === 'check') {
         event.returnValue = this.check(message.term)
       } else if (message.type === 'suggest') {
         event.returnValue = this.suggest(message.term)
+      } else if (message.type === 'add') {
+        event.returnValue = this.add(message.term)
       }
     })
 
@@ -55,14 +63,20 @@ class DictionaryProvider extends EventEmitter {
     })
 
     // Afterwards, set the timeout for loading the dictionaries
-    setTimeout(() => { this.reload() }, 5000)
+    setTimeout(() => {
+      this.reload()
+      this._loadUserDict() // On first start, load the user dictionary as well
+    }, 5000)
   }
 
   /**
    * Shuts down the provider
    * @return {Boolean} Whether or not the shutdown was successful
    */
-  shutdown () { return true }
+  shutdown () {
+    fs.writeFileSync(this._userDictionaryPath, this._userDictionary.join('\n'), 'utf8')
+    return true
+  }
 
   /**
    * (Re)Loads the dictionaries efficiently based upon the selected dictionaries
@@ -119,6 +133,24 @@ class DictionaryProvider extends EventEmitter {
   }
 
   /**
+   * Loads the user dictionary from file.
+   * @return {Promise} Will resolve after the dictionary has been loaded.
+   */
+  async _loadUserDict () {
+    try {
+      fs.lstatSync(this._userDictionaryPath)
+    } catch (e) {
+      // Create a new file (and add Zettlr as a correct word :3)
+      fs.writeFileSync(this._userDictionaryPath, 'Zettlr', 'utf8')
+    }
+    this._userDictionary = await readFile(this._userDictionaryPath, 'utf8')
+    this._userDictionary = this._userDictionary.split(/\n/)
+    // If the user dictionary is empty, the split will not create an array
+    // Send an invalidation message to the renderer so that it reloads all words
+    global.ipc.send('invalidate-dict')
+  }
+
+  /**
    * Checks the given term against the dictionaries and determines whether its
    * accurate.
    * @param  {String} term The word/term to check
@@ -141,6 +173,9 @@ class DictionaryProvider extends EventEmitter {
       if (typo.correct(term)) correct = true
     }
 
+    // Last but not least check the user dictionary
+    if (this._userDictionary.includes(term)) correct = true
+
     return correct
   }
 
@@ -160,6 +195,23 @@ class DictionaryProvider extends EventEmitter {
     }
 
     return suggestions
+  }
+
+  /**
+   * Adds the given term to the user dictionary
+   * @param {String} term The term to add
+   */
+  add (term) {
+    // Adds the given term to the user dictionary
+    if (!this._userDictionary.includes(term)) {
+      this._userDictionary.push(term)
+      console.log(this._userDictionary)
+      // Send an invalidation message to the renderer so that it reloads all words
+      global.ipc.send('invalidate-dict')
+    }
+
+    // Always return true (it'll get send to the renderer, but it won't deal with it.)
+    return true
   }
 
   /**